# services/tencent/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Instance API
    version: 0.0.1
paths:
    /api/images:
        get:
            tags:
                - Instance
            description: 选择镜像
            operationId: Instance_ListImages
            parameters:
                - name: secretId
                  in: query
                  schema:
                    type: string
                - name: secretKey
                  in: query
                  schema:
                    type: string
                - name: region
                  in: query
                  description: 地域参数，用来标识希望操作哪个地域的数据
                  schema:
                    type: string
                - name: offset
                  in: query
                  description: 偏移量，默认为0
                  schema:
                    type: integer
                    format: uint64
                - name: limit
                  in: query
                  description: 数量限制，默认为20，最大值为100
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListImagesResponse'
    /api/instance-types:
        get:
            tags:
                - Instance
            description: 选择实例规格
            operationId: Instance_ListInstanceTypes
            parameters:
                - name: secretId
                  in: query
                  schema:
                    type: string
                - name: secretKey
                  in: query
                  schema:
                    type: string
                - name: region
                  in: query
                  description: 地域参数，用来标识希望操作哪个地域的数据
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListInstanceTypesResponse'
    /api/instances:
        get:
            tags:
                - Instance
            description: 实例列表
            operationId: Instance_ListInstances
            parameters:
                - name: secretId
                  in: query
                  schema:
                    type: string
                - name: secretKey
                  in: query
                  schema:
                    type: string
                - name: region
                  in: query
                  description: 地域参数，用来标识希望操作哪个地域的数据
                  schema:
                    type: string
                - name: offset
                  in: query
                  description: 偏移量，默认为0
                  schema:
                    type: integer
                    format: int64
                - name: limit
                  in: query
                  description: 返回数量，默认为20，最大值为100。
                  schema:
                    type: integer
                    format: int64
                - name: instanceName
                  in: query
                  description: 按照【实例名称】进行过滤
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListInstancesResponse'
        post:
            tags:
                - Instance
            description: 创建实例
            operationId: Instance_CreateInstances
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateInstancesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateInstancesResponse'
    /api/instances/reboot:
        put:
            tags:
                - Instance
            description: 重启
            operationId: Instance_RebootInstances
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RebootInstancesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RebootInstancesResponse'
    /api/instances/start:
        put:
            tags:
                - Instance
            description: 开机
            operationId: Instance_StartInstances
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StartInstancesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StartInstancesResponse'
    /api/instances/stop:
        put:
            tags:
                - Instance
            description: 关机
            operationId: Instance_StopInstances
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StopInstancesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StopInstancesResponse'
    /api/instances/terminate:
        put:
            tags:
                - Instance
            description: 删除实例
            operationId: Instance_DeleteInstances
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteInstancesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteInstancesResponse'
    /api/regions:
        get:
            tags:
                - Instance
            description: 可使用的地域
            operationId: Instance_ListRegions
            parameters:
                - name: secretId
                  in: query
                  schema:
                    type: string
                - name: secretKey
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListRegionsResponse'
components:
    schemas:
        Any:
            type: object
            properties:
                typeUrl:
                    type: string
                    description: 'A URL/resource name that uniquely identifies the type of the serialized protocol buffer message. This string must contain at least one "/" character. The last segment of the URL''s path must represent the fully qualified name of the type (as in `path/google.protobuf.Duration`). The name should be in a canonical form (e.g., leading "." is not accepted). In practice, teams usually precompile into the binary all types that they expect it to use in the context of Any. However, for URLs which use the scheme `http`, `https`, or no scheme, one can optionally set up a type server that maps type URLs to message definitions as follows: * If no scheme is provided, `https` is assumed. * An HTTP GET on the URL must yield a [google.protobuf.Type][]   value in binary format, or produce an error. * Applications are allowed to cache lookup results based on the   URL, or have them precompiled into a binary to avoid any   lookup. Therefore, binary compatibility needs to be preserved   on changes to types. (Use versioned type names to manage   breaking changes.) Note: this functionality is not currently available in the official protobuf release, and it is not used for type URLs beginning with type.googleapis.com. Schemes other than `http`, `https` (or the empty scheme) might be used with implementation specific semantics.'
                value:
                    type: string
                    description: Must be a valid serialized protocol buffer of the above specified type.
                    format: bytes
            description: '`Any` contains an arbitrary serialized protocol buffer message along with a URL that describes the type of the serialized message. Protobuf library provides support to pack/unpack Any values in the form of utility functions or additional generated methods of the Any type. Example 1: Pack and unpack a message in C++.     Foo foo = ...;     Any any;     any.PackFrom(foo);     ...     if (any.UnpackTo(&foo)) {       ...     } Example 2: Pack and unpack a message in Java.     Foo foo = ...;     Any any = Any.pack(foo);     ...     if (any.is(Foo.class)) {       foo = any.unpack(Foo.class);     } Example 3: Pack and unpack a message in Python.     foo = Foo(...)     any = Any()     any.Pack(foo)     ...     if any.Is(Foo.DESCRIPTOR):       any.Unpack(foo)       ... Example 4: Pack and unpack a message in Go      foo := &pb.Foo{...}      any, err := anypb.New(foo)      if err != nil {        ...      }      ...      foo := &pb.Foo{}      if err := any.UnmarshalTo(foo); err != nil {        ...      } The pack methods provided by protobuf library will by default use ''type.googleapis.com/full.type.name'' as the type URL and the unpack methods only use the fully qualified type name after the last ''/'' in the type URL, for example "foo.bar.com/x/y.z" will yield type name "y.z". JSON The JSON representation of an `Any` value uses the regular representation of the deserialized, embedded message, with an additional field `@type` which contains the type URL. Example:     package google.profile;     message Person {       string first_name = 1;       string last_name = 2;     }     {       "@type": "type.googleapis.com/google.profile.Person",       "firstName": <string>,       "lastName": <string>     } If the embedded message type is well-known and has a custom JSON representation, that representation will be embedded adding a field `value` which holds the custom JSON in addition to the `@type` field. Example (for message [google.protobuf.Duration][]):     {       "@type": "type.googleapis.com/google.protobuf.Duration",       "value": "1.212s"     }'
        CreateInstancesRequest:
            type: object
            properties:
                secretId:
                    type: string
                secretKey:
                    type: string
                region:
                    type: string
                    description: 地域参数，用来标识希望操作哪个地域的数据
                imageId:
                    type: string
                    description: 镜像id
                name:
                    type: string
                    description: 实例名称, 不指定实例显示名称则默认显示‘未命名’。 购买多台实例，如果指定模式串{R:x}，表示生成数字[x, x+n-1]，其中n表示购买实例的数量，例如server_{R:3}，购买1台时，实例显示名称为server_3；购买2台时，实例显示名称分别为server_3，server_4。支持指定多个模式串{R:x}。 购买多台实例，如果不指定模式串，则在实例显示名称添加后缀1、2...n，其中n表示购买实例的数量，例如server_，购买2台时，实例显示名称分别为server_1，server_2。 最多支持60个字符（包含模式串）。
                instanceType:
                    type: string
                    description: 实例规格
                systemDisk:
                    $ref: '#/components/schemas/SystemDisk'
                instanceCount:
                    type: integer
                    description: 购买实例数量。包年包月实例取值范围：[1，300]，按量计费实例取值范围：[1，100]。默认取值：1。指定购买实例的数量不能超过用户所能购买的剩余配额数量，具体配额相关限制详见[CVM实例购买限制](https://cloud.tencent.com/document/product/213/2664)。
                    format: int64
                password:
                    type: string
                    description: '实例密码, Linux实例密码必须8到30位，至少包括两项[a-z]，[A-Z]、[0-9] 和 [( ) ` ~ ! @ # $ % ^ & * - + = | { } [ ] : ; '' , . ? / ]中的特殊符号。Windows实例密码必须12到30位，至少包括三项[a-z]，[A-Z]，[0-9] 和 [( ) ` ~ ! @ # $ % ^ & * - + = | { } [ ] : ; '' , . ? /]中的特殊符号。若不指定该参数，则由系统随机生成密码，并通过站内信方式通知到用户。'
                dryRun:
                    type: boolean
                    description: 是否只预检此次请求。 true：发送检查请求，不会创建实例。检查项包括是否填写了必需参数，请求格式，业务限制和云服务器库存。 如果检查不通过，则返回对应错误码； 如果检查通过，则返回RequestId. false（默认）：发送正常请求，通过检查后直接创建实例
        CreateInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        DeleteInstancesRequest:
            type: object
            properties:
                secretId:
                    type: string
                secretKey:
                    type: string
                region:
                    type: string
                    description: 地域参数，用来标识希望操作哪个地域的数据
                instanceIds:
                    type: array
                    items:
                        type: string
                    description: 一个或多个待操作的实例ID
        DeleteInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        ListImagesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        ListInstanceTypesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        ListInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        ListRegionsResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        RebootInstancesRequest:
            type: object
            properties:
                secretId:
                    type: string
                secretKey:
                    type: string
                region:
                    type: string
                    description: 地域参数，用来标识希望操作哪个地域的数据
                instanceIds:
                    type: array
                    items:
                        type: string
                    description: 一个或多个待操作的实例ID
                stopType:
                    type: string
                    description: 关机类型。 取值范围：  SOFT_FIRST：表示在正常关闭失败后进行强制关闭  HARD：直接强制关闭  SOFT：仅软关机 默认取值：SOFT。
        RebootInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        StartInstancesRequest:
            type: object
            properties:
                secretId:
                    type: string
                secretKey:
                    type: string
                region:
                    type: string
                    description: 地域参数，用来标识希望操作哪个地域的数据
                instanceIds:
                    type: array
                    items:
                        type: string
                    description: 一个或多个待操作的实例ID
        StartInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        StopInstancesRequest:
            type: object
            properties:
                secretId:
                    type: string
                secretKey:
                    type: string
                region:
                    type: string
                    description: 地域参数，用来标识希望操作哪个地域的数据
                instanceIds:
                    type: array
                    items:
                        type: string
                    description: 一个或多个待操作的实例ID
                stopType:
                    type: string
                    description: 实例的关闭模式。 取值范围：  SOFT_FIRST：表示在正常关闭失败后进行强制关闭  HARD：直接强制关闭  SOFT：仅软关机 默认取值：SOFT。
                stoppedMode:
                    type: string
                    description: 按量计费实例关机收费模式。 取值范围：  KEEP_CHARGING：关机继续收费  STOP_CHARGING：关机停止收费 默认取值：KEEP_CHARGING。 该参数只针对部分按量计费云硬盘实例生效，详情参考[按量计费实例关机不收费说明](https://cloud.tencent.com/document/product/213/19918)
        StopInstancesResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Any'
        SystemDisk:
            type: object
            properties:
                diskType:
                    type: string
                    description: 系统盘类型。系统盘类型限制详见[存储概述](https://cloud.tencent.com/document/product/213/4952)。 取值范围：  LOCAL_BASIC：本地硬盘  LOCAL_SSD：本地SSD硬盘  CLOUD_BASIC：普通云硬盘  CLOUD_SSD：SSD云硬盘  CLOUD_PREMIUM：高性能云硬盘 默认取值：当前有库存的硬盘类型。
                diskSize:
                    type: integer
                    description: 系统盘大小，单位：GB。默认值为 50
                    format: int64
                cdcId:
                    type: string
                    description: 所属的独享集群ID。
tags:
    - name: Instance